CryptoAPI/requirements.txt
uvicorn
fastapi
fastapi_simple_rate_limiter
pytz
sqlalchemy
PyJWT
requests
python-coinmarketcap
requests-cache
httpx
boto3
python-multipart